requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain_experimental
pandas
streamlit
tabulate
langchain-openai